+++ src/express/handlers/index.js
@@ -1,0 +1,2 @@
+export { default as ErrorHandler } from './error-handler'
+export { default as UserHandler } from './user-handler'
