+++ src/express/handlers/user-handler.js
@@ -1,21 +1,21 @@
 import { override } from 'core-decorators'
-import { User } from '../core'
-import { Repository } from '../sequelize'
-import HeaderValidationError from './errors/header-validation-error'
-import Resource from './resource'
+import { User } from '../../core'
+import { Repository } from '../../sequelize'
+import HeaderValidationError from '../errors/header-validation-error'
+import EntityHandler from './entity-handler'
 
 const B64_REGEX = /^([A-Za-z0-9+/]{4})*([A-Za-z0-9+/]{4}|[A-Za-z0-9+/]{3}=|[A-Za-z0-9+/]{2}==)$/
 
-export default class UserResource extends Resource(User) {
+export default class UserHandler extends EntityHandler(User) {
   static retrievePassword (req, res, next) {
-    let b64Password = req.get('hash')
+    let b64Password = req.get('encp')
 
     if (b64Password) {
       try {
         if (!B64_REGEX.test(b64Password)) throw new TypeError('Not a base64 string')
         req.password = Buffer.from(b64Password, 'base64').toString()
       } catch (err) {
-        return next(new HeaderValidationError('hash', 'Invalid hash format. Are you sure it is base64 encoded?'), err)
+        return next(new HeaderValidationError('encp', 'Invalid encp format. Are you sure it is base64 encoded?'), err)
       }
     }
 
@@ -30,7 +30,7 @@ export default class UserResource extends Resource(User) {
   @override
   static create (req, res, next) {
     if (!req.password) {
-      return next(new HeaderValidationError('hash', 'Hash must be informed'))
+      return next(new HeaderValidationError('encp', 'encp header must be informed'))
     }
 
     let user = new User(req.body)
