+++ src/express/index.js
@@ -1,0 +1,11 @@
+import express from 'express'
+import morgan from 'morgan'
+import bodyParser from 'body-parser'
+import { ErrorHandler, UserHandler } from './handlers'
+
+export default express()
+  .use(morgan(process.env.NODE_ENV === 'production' ? 'combined' : 'dev'))
+  .use(bodyParser.json({ type: 'application/json' }))
+  .use('/users', UserHandler.getRouter())
+  .use(ErrorHandler.getRouter())
+  .get('/', (req, res) => res.send('Hello World! \\o/'))
