--- src/resources/index.js
@@ -1,1 +1,0 @@
-export { default as UserResource } from './user-resource'
